added question helper to navigate to next cursor after updating api
