+++ src/main/features/rfi/controller/questions.ts
@@ -3,6 +3,7 @@ import {operations } from '../../../utils/operations/operations';
 import { DynamicFrameworkInstance } from '../util/fetch/dyanmicframeworkInstance';
 import { ObjectModifiers } from '../util/operations/objectremoveEmptyString';
 import { ErrorView } from '../../../common/shared/error/errorView';
+import {QuestionHelper} from '../helpers/question'
 
 /**
  * @Controller
@@ -110,7 +111,7 @@ export var array : any = [];
                      let answerBaseURL =  `/tenders/projects/${proc_id}/events/${event_id}/criteria/${id}/groups/${group_id}/questions/${iteration.questionNo}`;
                      let postData = await DynamicFrameworkInstance.Instance(SESSION_ID).put(answerBaseURL, answerBody);
                      console.log(postData)
-                     POST_UPDATINGDATA(proc_id, event_id, SESSION_ID, group_id, agreement_id, id, res)
+                     QuestionHelper.AFTER_UPDATINGDATA(ErrorView,DynamicFrameworkInstance,proc_id, event_id, SESSION_ID, group_id, agreement_id, id, res);
                    } catch (error) {
                       res.redirect('/404')
                    }
@@ -130,7 +131,7 @@ export var array : any = [];
                let answerBaseURL =  `/tenders/projects/${proc_id}/events/${event_id}/criteria/${id}/groups/${group_id}/questions/${question_id}`;
                let postData = await DynamicFrameworkInstance.Instance(SESSION_ID).put(answerBaseURL, answerBody);
                console.log(postData)
-                  POST_UPDATINGDATA(proc_id, event_id, SESSION_ID, group_id, agreement_id, id, res)
+               QuestionHelper.AFTER_UPDATINGDATA(ErrorView,DynamicFrameworkInstance,proc_id, event_id, SESSION_ID, group_id, agreement_id, id, res);
              } catch (error) {
                 res.redirect('/404')
              }
@@ -149,55 +150,4 @@ export var array : any = [];
 
 
 
- const POST_UPDATINGDATA = async (proc_id : any, event_id: any, SESSION_ID: string, group_id: any, agreement_id: any, id: any, res: express.Response) => {
-      /**
-             * @Path
-             * @Next
-             * Sorting and following to the next path
-             */
-
-       let baseURL : any = `/tenders/projects/${proc_id}/events/${event_id}/criteria`;
-       try {
-          let fetch_dynamic_api = await DynamicFrameworkInstance.Instance(SESSION_ID).get(baseURL);
-          let fetch_dynamic_api_data = fetch_dynamic_api?.data; 
-          let extracted_criterion_based = fetch_dynamic_api_data?.map((criterian : any)=> criterian?.id);
-          var criterianStorage: any = [];
-          for (let aURI of extracted_criterion_based) {
-             let criterian_bas_url =  `/tenders/projects/${proc_id}/events/${event_id}/criteria/${aURI}/groups`;
-             let fetch_criterian_group_data = await DynamicFrameworkInstance.Instance(SESSION_ID).get(criterian_bas_url);
-             let criterian_array = fetch_criterian_group_data?.data;
-             let rebased_object_with_requirements = criterian_array?.map((anItem: any)=> {
-                let object = anItem;
-                object['criterianId'] = aURI;
-                return object;
-             })
-             criterianStorage.push(rebased_object_with_requirements)
-          }
-          criterianStorage = criterianStorage.flat();
-          let sorted_ascendingly = criterianStorage.map((aCriterian : any) => {
-             let object = aCriterian;
-             object.OCDS['id'] = aCriterian.OCDS['id']?.split('Group ').join('');
-             return object;
-          }).sort((current_cursor: any, iterator_cursor : any) => Number(current_cursor.OCDS['id']) - Number(iterator_cursor.OCDS['id'])).map((aCriterian : any ) => {
-             var object = aCriterian;
-             object.OCDS['id'] = `Group ${aCriterian.OCDS['id']}`
-             return object;
-          });
-          let current_cursor = sorted_ascendingly?.findIndex((pointer : any) => pointer.OCDS['id'] === group_id);
-          let check_for_overflowing : Boolean = current_cursor < sorted_ascendingly.length;
-          if(check_for_overflowing){
-             let next_cursor = current_cursor + 1;
-             let next_cursor_object = sorted_ascendingly[next_cursor];
-             let next_group_id = next_cursor_object.OCDS['id'];
-             let next_criterian_id = next_cursor_object['criterianId'];
-             let base_url = `/rfi/questions?agreement_id=${agreement_id}&proc_id=${proc_id}&event_id=${event_id}&id=${next_criterian_id}&group_id=${next_group_id}`
-             console.log(id)
-             res.redirect(base_url)
-          }
-          else{
-             // do some logic here 
-          }
-       } catch (error) {
-          res.redirect(ErrorView.notfound)
-       }
- }
+
